Split CompanionServer: web pages out, route pipeline into named groups
CompanionServer.kt 5,755 -> 4,910 lines, and its 2,202-line
configurePipeline is now eleven named functions.

Two separate changes, both verbatim moves:

1. server/CompanionWebPages.kt takes the five self-contained pages the server
   serves to phones -- Q&A submission, voting and moderation, and the
   presentation remote. 901 lines of static markup and CSS with three Kotlin
   interpolations in the whole set, which is a fifth of the file and is not
   server logic. Editing a page no longer means scrolling through the request
   pipeline, and reading the pipeline no longer means scrolling through HTML.
   They compiled as top-level functions on the first attempt, which confirms
   they never touched class state.

2. configurePipeline's routing{} body is split into eleven member extension
   functions on Route, called in the order the routes were declared:
   certificateRoutes, infoAndSongRoutes, scheduleRoutes,
   bibleAndDictionaryRoutes, presentationRoutes, presentationRemoteRoutes,
   mediaAndAssetRoutes, webSocketRoute, lowerThirdAndAtemRoutes,
   browserSourceRoutes, qaRoutes.

   Order is load-bearing -- Ktor matches routes in declaration order, so
   overlapping paths would resolve differently if the sequence changed. It is
   preserved exactly, and verified: diffing the extracted bodies against
   CompanionServer.kt at HEAD comes back identical.

   They stay member extensions in the same file rather than moving out.
   Relocating them would mean widening dozens of private fields to internal,
   which is a far larger blast radius than a navigability gain justifies.

The 14 server suites, which drive these routes over real HTTP and WebSocket,
pass unchanged -- that is the evidence the split preserved behaviour.
